[New Feature] Searchable Settings Menu
PR #2733 Implemented the ability to specify a search string in the settings menu
