+++ app/src/main/java/me/ccrama/redditslide/Activities/MainActivity.java
@@ -147,6 +147,8 @@ import me.ccrama.redditslide.Constants;
 import me.ccrama.redditslide.ContentType;
 import me.ccrama.redditslide.FDroid;
 import me.ccrama.redditslide.Fragments.CommentPage;
+import me.ccrama.redditslide.Fragments.SettingsGeneralFragment;
+import me.ccrama.redditslide.Fragments.SettingsThemeFragment;
 import me.ccrama.redditslide.Fragments.SubmissionsView;
 import me.ccrama.redditslide.HasSeen;
 import me.ccrama.redditslide.ImageFlairs;
@@ -1516,16 +1518,16 @@ public class MainActivity extends BaseActivity
             setToolbarClick();
         }*/
         //Only refresh the view if a Setting was altered
-        if (Settings.changed || SettingsTheme.changed) {
+        if (Settings.changed || SettingsThemeFragment.changed) {
 
             reloadSubs();
             //If the user changed a Setting regarding the app's theme, restartTheme()
-            if (SettingsTheme.changed /* todo maybe later || (usedArray != null && usedArray.size() != UserSubscriptions.getSubscriptions(this).size())*/) {
+            if (SettingsThemeFragment.changed /* todo maybe later || (usedArray != null && usedArray.size() != UserSubscriptions.getSubscriptions(this).size())*/) {
                 restartTheme();
             }
 
             //Need to change the subreddit search method
-            if (SettingsGeneral.searchChanged) {
+            if (SettingsGeneralFragment.searchChanged) {
                 setDrawerSubList();
 
                 if (SettingValues.subredditSearchMethod
@@ -1542,9 +1544,9 @@ public class MainActivity extends BaseActivity
                     setupSubredditSearchToolbar();
                     setDrawerSubList();
                 }
-                SettingsGeneral.searchChanged = false;
+                SettingsGeneralFragment.searchChanged = false;
             }
-            SettingsTheme.changed = false;
+            SettingsThemeFragment.changed = false;
             Settings.changed = false;
             setToolbarClick();
         }
