+++ app/src/main/java/me/ccrama/redditslide/Fragments/SettingsFragment.java
@@ -1,0 +1,9 @@
+package me.ccrama.redditslide.Fragments;
+
+public class SettingsFragment {
+
+    public interface RestartActivity {
+        void restartActivity();
+    }
+
+}
