+++ app/src/main/res/layout/activity_settings_datasaving.xml
@@ -5,106 +5,13 @@
     android:background="?attr/card_background">
 
     <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+        android:id="@+id/settings_datasaving"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:orientation="vertical">
 
         <include layout="@layout/settings_toolbar" />
 
-        <TextView
-            style="@style/TextAppearance.AppCompat.Body2"
-            android:id="@+id/dataSavingModeHeader"
-            android:text="@string/settings_section_data_saving_mode"
-            android:paddingStart="16dp"
-            android:paddingEnd="16dp"
-            android:paddingTop="16dp"
-            android:layout_width="match_parent"
-            android:textColor="?attr/colorAccent"
-            android:layout_height="wrap_content" />
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:orientation="vertical">
-
-            <RelativeLayout
-                android:id="@+id/datasavetype"
-                android:layout_width="match_parent"
-                android:layout_height="56dp"
-                android:background="?android:selectableItemBackground"
-                android:gravity="center_vertical"
-                android:orientation="horizontal"
-                android:paddingEnd="16dp"
-                android:paddingStart="16dp">
-
-                <LinearLayout
-                    android:layout_marginEnd="64dp"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:layout_centerVertical="true"
-                    android:orientation="vertical">
-
-                    <TextView
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:text="@string/settings_data_enable"
-                        android:textColor="?attr/fontColor"
-                        android:textSize="14sp" />
-
-                    <TextView
-                        android:id="@+id/lowquality"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:alpha=".86"
-                        android:textColor="?attr/fontColor"
-                        android:textSize="13sp" />
-                </LinearLayout>
-            </RelativeLayout>
-
-            <TextView
-                style="@style/TextAppearance.AppCompat.Body2"
-                android:id="@+id/imageQualityHeader"
-                android:text="@string/settings_section_image_quality"
-                android:paddingStart="16dp"
-                android:paddingEnd="16dp"
-                android:paddingTop="16dp"
-                android:layout_width="match_parent"
-                android:textColor="?attr/colorAccent"
-                android:layout_height="wrap_content" />
-
-            <RelativeLayout
-                    android:id="@+id/datasavequality"
-                    android:layout_width="match_parent"
-                    android:layout_height="56dp"
-                    android:background="?android:selectableItemBackground"
-                    android:gravity="center_vertical"
-                    android:orientation="horizontal"
-                    android:paddingEnd="16dp"
-                    android:paddingStart="16dp">
-
-                <LinearLayout
-                        android:layout_marginEnd="64dp"
-                        android:layout_width="match_parent"
-                        android:layout_height="wrap_content"
-                        android:layout_centerVertical="true"
-                        android:orientation="vertical">
-
-                    <TextView
-                            android:layout_width="wrap_content"
-                            android:layout_height="wrap_content"
-                            android:text="Image quality when Data Saving enabled"
-                            android:textColor="?attr/fontColor"
-                            android:textSize="14sp" />
-
-                    <TextView
-                            android:id="@+id/currentmode"
-                            android:layout_width="wrap_content"
-                            android:layout_height="wrap_content"
-                            android:alpha=".86"
-                            android:textColor="?attr/fontColor"
-                            android:textSize="13sp" />
-                </LinearLayout>
-            </RelativeLayout>
-        </LinearLayout>
     </LinearLayout>
+
 </ScrollView>
